+++ plover/scripts/main.py
@@ -115,8 +115,11 @@ def main():
                         code = entrypoint.load()()
                     except SystemExit as e:
                         code = e.code
-                if code is None:
-                    code = 0
+                    if code is None:
+                        code = 0
+                    elif not isinstance(code, int):  # sys.exit() also takes a string
+                        log.error(code)
+                        code = 1
             else:
                 print("available script(s):")
                 dist = None
@@ -163,7 +166,7 @@ def main():
                     code = -1
                 else:
                     code = 0
-    except:
+    except Exception:
         gui.show_error("Unexpected error", traceback.format_exc())
         code = 2
     # Execute atexit handlers.